Feat : return proper status code

+++ src/main/java/com/example/wantedpreonboardingbackend/jobposting/JobPostingController.java
@@ -1,7 +1,9 @@
 package com.example.wantedpreonboardingbackend.jobposting;
 
 import com.example.wantedpreonboardingbackend.company.CompanyRepository;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.AllArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -15,21 +17,45 @@ public class JobPostingController {
     CompanyRepository companyRepository;
 
     @PostMapping("/jobposting")
-    public void createJobPosting(@RequestBody() JobPostingRequestDto jobPostingRequestDto) {
-        JobPosting jobPosting = jobPostingRequestDto.toEntity(companyRepository);
-        jobPosting.getCompany().getJobPostings().add(jobPosting);
-        repository.save(jobPosting);
+    public ResponseEntity createJobPosting(@RequestBody() JobPostingRequestDto jobPostingRequestDto) {
+        try {
+            JobPosting jobPosting = jobPostingRequestDto.toEntity(companyRepository);
+            jobPosting.getCompany().getJobPostings().add(jobPosting);
+            repository.save(jobPosting);
+
+            return ResponseEntity.created(null).build();
+        }
+        catch (Exception e) {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PutMapping("/jobposting/{id}")
-    public void updateJobPosting(@PathVariable int id, @RequestBody JobPostingRequestDto jobPostingRequestDto) {
-        JobPosting newJobPosting = jobPostingRequestDto.updateAndToEntity(repository.findById(id).orElse(null));
-        repository.save(newJobPosting);
+    public ResponseEntity updateJobPosting(@PathVariable int id, @RequestBody JobPostingRequestDto jobPostingRequestDto) {
+        try {
+            JobPosting newJobPosting = jobPostingRequestDto.updateAndToEntity(repository.findById(id).orElseThrow(() -> new Exception("No JobPosting!")));
+            repository.save(newJobPosting);
+
+            return ResponseEntity.created(null).build();
+        }
+        catch (Exception e) {
+            return ResponseEntity.notFound().build();
+        }
+
     }
 
     @DeleteMapping("/jobposting/{id}")
-    public void deleteJobPosting(@PathVariable int id) {
-        repository.deleteById(id);
+    public ResponseEntity deleteJobPosting(@PathVariable int id) {
+        JobPosting jobPosting = repository.findById(id).orElse(null);
+        if (jobPosting == null) {
+            return ResponseEntity.notFound().build();
+        }
+        else {
+            jobPosting.getCompany().getJobPostings().remove(jobPosting);
+            repository.deleteById(id);
+            return ResponseEntity.noContent().build();
+
+        }
     }
 
     @GetMapping("/jobposting")
@@ -62,12 +88,17 @@ public class JobPostingController {
     }
 
     @GetMapping("/jobposting/{id}")
-    public JobPostingDetailResponseDto retrieveJobPostingDetail(@PathVariable int id) {
-        JobPosting jobPosting = repository.findById(id).orElse(null);
-        JobPostingDetailResponseDto jobPostingDetailResponseDto = new JobPostingDetailResponseDto();
-        jobPostingDetailResponseDto.fromEntity(jobPosting);
+    public ResponseEntity retrieveJobPostingDetail(@PathVariable int id) {
+        try {
+            JobPosting jobPosting = repository.findById(id).orElseThrow(() -> new Exception("No JobPosting!"));
+            JobPostingDetailResponseDto jobPostingDetailResponseDto = new JobPostingDetailResponseDto();
+            jobPostingDetailResponseDto.fromEntity(jobPosting);
 
-        return jobPostingDetailResponseDto;
+            return ResponseEntity.ok(jobPostingDetailResponseDto);
+        }
+        catch (Exception e) {
+            return ResponseEntity.notFound().build();
+        }
     }
